Move the new test into its own file
