+++ manifests/golang.yml
@@ -517,6 +517,8 @@ tests/:
       Test_AWS_API_Gateway_Inferred_Span_Creation: missing_feature
     test_otel_drop_in.py:
       Test_Otel_Drop_In: missing_feature
+  otel/:
+    test_context_propagation.py:
       Test_Otel_Drop_In_Default_Propagator:
         '*': irrelevant
         net-http: v1.70.1
