+++ manifests/java.yml
@@ -1595,11 +1595,13 @@ tests/:
       Test_Otel_Drop_In:
         '*': missing_feature
         spring-boot: v1.39.0
+    test_sql.py:
+      Test_Sql: bug (APMAPI-729)
+  otel/:
+    test_context_propagation.py:
       Test_Otel_Drop_In_Default_Propagator:
         '*': missing_feature
         spring-boot: v1.39.0
-    test_sql.py:
-      Test_Sql: bug (APMAPI-729)
   parametric/:
     test_config_consistency.py:
       Test_Config_Dogstatsd: missing_feature (default hostname is inconsistent)
