+++ manifests/php.yml
@@ -382,6 +382,8 @@ tests/:
       Test_AWS_API_Gateway_Inferred_Span_Creation: missing_feature
     test_otel_drop_in.py:
       Test_Otel_Drop_In: missing_feature
+  otel/:
+    test_context_propagation.py:
       Test_Otel_Drop_In_Default_Propagator: incomplete_test_app (endpoint not implemented)
   parametric/:
     test_128_bit_traceids.py:
