+++ manifests/python.yml
@@ -760,6 +760,8 @@ tests/:
       Test_AWS_API_Gateway_Inferred_Span_Creation: missing_feature
     test_otel_drop_in.py:
       Test_Otel_Drop_In: missing_feature
+  otel/:
+    test_context_propagation.py:
       Test_Otel_Drop_In_Default_Propagator:
         '*': irrelevant
         flask-poc: v2.19.0
