+++ manifests/ruby.yml
@@ -400,14 +400,16 @@ tests/:
       Test_AWS_API_Gateway_Inferred_Span_Creation: missing_feature
     test_otel_drop_in.py:
       Test_Otel_Drop_In: missing_feature
-      Test_Otel_Drop_In_Default_Propagator:
-        '*': irrelevant (endpoint not implemented)
-        rails70: v2.0.0
   k8s_lib_injection/:
     test_k8s_lib_injection_profiling.py:
       TestK8sLibInjectioProfilingClusterEnabled: missing_feature
       TestK8sLibInjectioProfilingClusterOverride: missing_feature
       TestK8sLibInjectioProfilingDisabledByDefault: missing_feature
+  otel/:
+    test_context_propagation.py:
+      Test_Otel_Drop_In_Default_Propagator:
+        '*': irrelevant (endpoint not implemented)
+        rails70: v2.0.0
   parametric/:
     test_config_consistency.py:
       Test_Config_Dogstatsd: missing_feature
